+++ ts-client/src/dealMatcherClient/indexerClient/generated.types.ts
@@ -55,6 +55,7 @@ export type CapacityCommitment = {
   nextCCFailedEpoch: Scalars['BigInt']['output'];
   peer: Peer;
   provider: Provider;
+  /** This field represents Ratio [0, 1] only when it is divided by PRECISION constant of Core contract. */
   rewardDelegatorRate: Scalars['Int']['output'];
   rewardWithdrawn: Scalars['BigInt']['output'];
   snapshotEpoch: Scalars['BigInt']['output'];
@@ -98,13 +99,13 @@ export type CapacityCommitmentSubmittedProofsArgs = {
 export type CapacityCommitmentStatsPerEpoch = {
   __typename?: 'CapacityCommitmentStatsPerEpoch';
   activeUnitCount: Scalars['Int']['output'];
-  blockNumberEnd: Scalars['BigInt']['output'];
-  blockNumberStart: Scalars['BigInt']['output'];
   capacityCommitment: CapacityCommitment;
   /** Should be update after ComputeUnitPerEpochStat.submittedProofsCount reached min proofs border. */
   computeUnitsWithMinRequiredProofsSubmittedCounter: Scalars['Int']['output'];
   currentCCNextCCFailedEpoch: Scalars['BigInt']['output'];
+  /** Additional field to support ordering by epoch. */
   epoch: Scalars['BigInt']['output'];
+  epochStatistic: EpochStatistic;
   exitedUnitCount: Scalars['Int']['output'];
   id: Scalars['ID']['output'];
   nextAdditionalActiveUnitCount: Scalars['Int']['output'];
@@ -135,22 +136,6 @@ export type CapacityCommitmentStatsPerEpoch_Filter = {
   activeUnitCount_not?: InputMaybe<Scalars['Int']['input']>;
   activeUnitCount_not_in?: InputMaybe<Array<Scalars['Int']['input']>>;
   and?: InputMaybe<Array<InputMaybe<CapacityCommitmentStatsPerEpoch_Filter>>>;
-  blockNumberEnd?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_gt?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_gte?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
-  blockNumberEnd_lt?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_lte?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_not?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberEnd_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
-  blockNumberStart?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_gt?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_gte?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
-  blockNumberStart_lt?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_lte?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_not?: InputMaybe<Scalars['BigInt']['input']>;
-  blockNumberStart_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
   capacityCommitment?: InputMaybe<Scalars['String']['input']>;
   capacityCommitment_?: InputMaybe<CapacityCommitment_Filter>;
   capacityCommitment_contains?: InputMaybe<Scalars['String']['input']>;
@@ -189,6 +174,27 @@ export type CapacityCommitmentStatsPerEpoch_Filter = {
   currentCCNextCCFailedEpoch_not?: InputMaybe<Scalars['BigInt']['input']>;
   currentCCNextCCFailedEpoch_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
   epoch?: InputMaybe<Scalars['BigInt']['input']>;
+  epochStatistic?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_?: InputMaybe<EpochStatistic_Filter>;
+  epochStatistic_contains?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_contains_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_ends_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_ends_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_gt?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_gte?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_in?: InputMaybe<Array<Scalars['String']['input']>>;
+  epochStatistic_lt?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_lte?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_contains?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_contains_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_ends_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_ends_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_in?: InputMaybe<Array<Scalars['String']['input']>>;
+  epochStatistic_not_starts_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_starts_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
   epoch_gt?: InputMaybe<Scalars['BigInt']['input']>;
   epoch_gte?: InputMaybe<Scalars['BigInt']['input']>;
   epoch_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
@@ -242,8 +248,6 @@ export type CapacityCommitmentStatsPerEpoch_Filter = {
 
 export type CapacityCommitmentStatsPerEpoch_OrderBy =
   | 'activeUnitCount'
-  | 'blockNumberEnd'
-  | 'blockNumberStart'
   | 'capacityCommitment'
   | 'capacityCommitment__activeUnitCount'
   | 'capacityCommitment__collateralPerUnit'
@@ -269,6 +273,12 @@ export type CapacityCommitmentStatsPerEpoch_OrderBy =
   | 'computeUnitsWithMinRequiredProofsSubmittedCounter'
   | 'currentCCNextCCFailedEpoch'
   | 'epoch'
+  | 'epochStatistic'
+  | 'epochStatistic__endBlock'
+  | 'epochStatistic__endTimestamp'
+  | 'epochStatistic__id'
+  | 'epochStatistic__startBlock'
+  | 'epochStatistic__startTimestamp'
   | 'exitedUnitCount'
   | 'id'
   | 'nextAdditionalActiveUnitCount'
@@ -373,6 +383,7 @@ export type CapacityCommitmentToComputeUnit_OrderBy =
   | 'capacityCommitment__totalFailCount'
   | 'computeUnit'
   | 'computeUnit__createdAt'
+  | 'computeUnit__deleted'
   | 'computeUnit__id'
   | 'computeUnit__submittedProofsCount'
   | 'computeUnit__workerId'
@@ -622,6 +633,7 @@ export type CapacityCommitment_OrderBy =
   | 'peer__currentCCCollateralDepositedAt'
   | 'peer__currentCCEndEpoch'
   | 'peer__currentCCNextCCFailedEpoch'
+  | 'peer__deleted'
   | 'peer__id'
   | 'peer__isAnyJoinedDeals'
   | 'provider'
@@ -647,6 +659,7 @@ export type ComputeUnit = {
   __typename?: 'ComputeUnit';
   createdAt: Scalars['BigInt']['output'];
   deal?: Maybe<Deal>;
+  deleted: Scalars['Boolean']['output'];
   id: Scalars['ID']['output'];
   peer: Peer;
   /** In order to simplify relation for query. */
@@ -670,7 +683,7 @@ export type ComputeUnitPerEpochStat = {
   __typename?: 'ComputeUnitPerEpochStat';
   capacityCommitment?: Maybe<CapacityCommitment>;
   computeUnit: ComputeUnit;
-  epoch: Scalars['BigInt']['output'];
+  epochStatistic: EpochStatistic;
   id: Scalars['ID']['output'];
   submittedProofsCount: Scalars['Int']['output'];
 };
@@ -721,14 +734,27 @@ export type ComputeUnitPerEpochStat_Filter = {
   computeUnit_not_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
   computeUnit_starts_with?: InputMaybe<Scalars['String']['input']>;
   computeUnit_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
-  epoch?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_gt?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_gte?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
-  epoch_lt?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_lte?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_not?: InputMaybe<Scalars['BigInt']['input']>;
-  epoch_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  epochStatistic?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_?: InputMaybe<EpochStatistic_Filter>;
+  epochStatistic_contains?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_contains_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_ends_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_ends_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_gt?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_gte?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_in?: InputMaybe<Array<Scalars['String']['input']>>;
+  epochStatistic_lt?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_lte?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_contains?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_contains_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_ends_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_ends_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_in?: InputMaybe<Array<Scalars['String']['input']>>;
+  epochStatistic_not_starts_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_not_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_starts_with?: InputMaybe<Scalars['String']['input']>;
+  epochStatistic_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
   id?: InputMaybe<Scalars['ID']['input']>;
   id_gt?: InputMaybe<Scalars['ID']['input']>;
   id_gte?: InputMaybe<Scalars['ID']['input']>;
@@ -773,10 +799,16 @@ export type ComputeUnitPerEpochStat_OrderBy =
   | 'capacityCommitment__totalFailCount'
   | 'computeUnit'
   | 'computeUnit__createdAt'
+  | 'computeUnit__deleted'
   | 'computeUnit__id'
   | 'computeUnit__submittedProofsCount'
   | 'computeUnit__workerId'
-  | 'epoch'
+  | 'epochStatistic'
+  | 'epochStatistic__endBlock'
+  | 'epochStatistic__endTimestamp'
+  | 'epochStatistic__id'
+  | 'epochStatistic__startBlock'
+  | 'epochStatistic__startTimestamp'
   | 'id'
   | 'submittedProofsCount';
 
@@ -813,6 +845,10 @@ export type ComputeUnit_Filter = {
   deal_not_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
   deal_starts_with?: InputMaybe<Scalars['String']['input']>;
   deal_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  deleted?: InputMaybe<Scalars['Boolean']['input']>;
+  deleted_in?: InputMaybe<Array<Scalars['Boolean']['input']>>;
+  deleted_not?: InputMaybe<Scalars['Boolean']['input']>;
+  deleted_not_in?: InputMaybe<Array<Scalars['Boolean']['input']>>;
   id?: InputMaybe<Scalars['ID']['input']>;
   id_gt?: InputMaybe<Scalars['ID']['input']>;
   id_gte?: InputMaybe<Scalars['ID']['input']>;
@@ -912,6 +948,7 @@ export type ComputeUnit_OrderBy =
   | 'deal__registeredWorkersCurrentCount'
   | 'deal__targetWorkers'
   | 'deal__withdrawalSum'
+  | 'deleted'
   | 'id'
   | 'peer'
   | 'peer__computeUnitsInDeal'
@@ -919,6 +956,7 @@ export type ComputeUnit_OrderBy =
   | 'peer__currentCCCollateralDepositedAt'
   | 'peer__currentCCEndEpoch'
   | 'peer__currentCCNextCCFailedEpoch'
+  | 'peer__deleted'
   | 'peer__id'
   | 'peer__isAnyJoinedDeals'
   | 'provider'
@@ -1204,6 +1242,7 @@ export type DealToJoinedOfferPeer_OrderBy =
   | 'peer__currentCCCollateralDepositedAt'
   | 'peer__currentCCEndEpoch'
   | 'peer__currentCCNextCCFailedEpoch'
+  | 'peer__deleted'
   | 'peer__id'
   | 'peer__isAnyJoinedDeals';
 
@@ -1305,6 +1344,7 @@ export type DealToPeer_OrderBy =
   | 'peer__currentCCCollateralDepositedAt'
   | 'peer__currentCCEndEpoch'
   | 'peer__currentCCNextCCFailedEpoch'
+  | 'peer__deleted'
   | 'peer__id'
   | 'peer__isAnyJoinedDeals';
 
@@ -1652,6 +1692,71 @@ export type Effector_OrderBy =
   | 'id'
   | 'offers';
 
+/** This model is designed to store epoch related information. Note, that in other models it is more efficient to store epoch as bigint rather than relation to that model (with relation you could complicate your queries by epoch, or you will need to store additional epoch field anyway). */
+export type EpochStatistic = {
+  __typename?: 'EpochStatistic';
+  endBlock: Scalars['BigInt']['output'];
+  endTimestamp: Scalars['BigInt']['output'];
+  /** Epoch number. Note, that for current epoch right boarder is approximate. */
+  id: Scalars['ID']['output'];
+  startBlock: Scalars['BigInt']['output'];
+  startTimestamp: Scalars['BigInt']['output'];
+};
+
+export type EpochStatistic_Filter = {
+  /** Filter for the block changed event. */
+  _change_block?: InputMaybe<BlockChangedFilter>;
+  and?: InputMaybe<Array<InputMaybe<EpochStatistic_Filter>>>;
+  endBlock?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_gt?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_gte?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  endBlock_lt?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_lte?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_not?: InputMaybe<Scalars['BigInt']['input']>;
+  endBlock_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  endTimestamp?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_gt?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_gte?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  endTimestamp_lt?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_lte?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_not?: InputMaybe<Scalars['BigInt']['input']>;
+  endTimestamp_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  id?: InputMaybe<Scalars['ID']['input']>;
+  id_gt?: InputMaybe<Scalars['ID']['input']>;
+  id_gte?: InputMaybe<Scalars['ID']['input']>;
+  id_in?: InputMaybe<Array<Scalars['ID']['input']>>;
+  id_lt?: InputMaybe<Scalars['ID']['input']>;
+  id_lte?: InputMaybe<Scalars['ID']['input']>;
+  id_not?: InputMaybe<Scalars['ID']['input']>;
+  id_not_in?: InputMaybe<Array<Scalars['ID']['input']>>;
+  or?: InputMaybe<Array<InputMaybe<EpochStatistic_Filter>>>;
+  startBlock?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_gt?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_gte?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  startBlock_lt?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_lte?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_not?: InputMaybe<Scalars['BigInt']['input']>;
+  startBlock_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  startTimestamp?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_gt?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_gte?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  startTimestamp_lt?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_lte?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_not?: InputMaybe<Scalars['BigInt']['input']>;
+  startTimestamp_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+};
+
+export type EpochStatistic_OrderBy =
+  | 'endBlock'
+  | 'endTimestamp'
+  | 'id'
+  | 'startBlock'
+  | 'startTimestamp';
+
 export type GraphNetwork = {
   __typename?: 'GraphNetwork';
   capacityCommitmentsTotal: Scalars['BigInt']['output'];
@@ -1659,6 +1764,7 @@ export type GraphNetwork = {
   capacityMaxFailedRatio?: Maybe<Scalars['Int']['output']>;
   coreContractAddress?: Maybe<Scalars['String']['output']>;
   coreEpochDuration?: Maybe<Scalars['Int']['output']>;
+  corePrecision?: Maybe<Scalars['Int']['output']>;
   dealsTotal: Scalars['BigInt']['output'];
   effectorsTotal: Scalars['BigInt']['output'];
   /** ID is set to 1 */
@@ -1668,6 +1774,9 @@ export type GraphNetwork = {
   minRequiredProofsPerEpoch?: Maybe<Scalars['Int']['output']>;
   offersTotal: Scalars['BigInt']['output'];
   proofsTotal: Scalars['BigInt']['output'];
+  /** Providers that register themselves in the network with setInfo() method. */
+  providersRegisteredTotal: Scalars['BigInt']['output'];
+  /** @deprecated TODO: deprecate because it is not used. changed to providersRegisteredTotal used. */
   providersTotal: Scalars['BigInt']['output'];
   tokensTotal: Scalars['BigInt']['output'];
 };
@@ -1740,6 +1849,14 @@ export type GraphNetwork_Filter = {
   coreEpochDuration_lte?: InputMaybe<Scalars['Int']['input']>;
   coreEpochDuration_not?: InputMaybe<Scalars['Int']['input']>;
   coreEpochDuration_not_in?: InputMaybe<Array<Scalars['Int']['input']>>;
+  corePrecision?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_gt?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_gte?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_in?: InputMaybe<Array<Scalars['Int']['input']>>;
+  corePrecision_lt?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_lte?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_not?: InputMaybe<Scalars['Int']['input']>;
+  corePrecision_not_in?: InputMaybe<Array<Scalars['Int']['input']>>;
   dealsTotal?: InputMaybe<Scalars['BigInt']['input']>;
   dealsTotal_gt?: InputMaybe<Scalars['BigInt']['input']>;
   dealsTotal_gte?: InputMaybe<Scalars['BigInt']['input']>;
@@ -1817,6 +1934,14 @@ export type GraphNetwork_Filter = {
   proofsTotal_lte?: InputMaybe<Scalars['BigInt']['input']>;
   proofsTotal_not?: InputMaybe<Scalars['BigInt']['input']>;
   proofsTotal_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  providersRegisteredTotal?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_gt?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_gte?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
+  providersRegisteredTotal_lt?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_lte?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_not?: InputMaybe<Scalars['BigInt']['input']>;
+  providersRegisteredTotal_not_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
   providersTotal?: InputMaybe<Scalars['BigInt']['input']>;
   providersTotal_gt?: InputMaybe<Scalars['BigInt']['input']>;
   providersTotal_gte?: InputMaybe<Scalars['BigInt']['input']>;
@@ -1841,6 +1966,7 @@ export type GraphNetwork_OrderBy =
   | 'capacityMaxFailedRatio'
   | 'coreContractAddress'
   | 'coreEpochDuration'
+  | 'corePrecision'
   | 'dealsTotal'
   | 'effectorsTotal'
   | 'id'
@@ -1849,11 +1975,13 @@ export type GraphNetwork_OrderBy =
   | 'minRequiredProofsPerEpoch'
   | 'offersTotal'
   | 'proofsTotal'
+  | 'providersRegisteredTotal'
   | 'providersTotal'
   | 'tokensTotal';
 
 export type Offer = {
   __typename?: 'Offer';
+  /** It depends on if CU in deal or not. */
   computeUnitsAvailable?: Maybe<Scalars['Int']['output']>;
   computeUnitsTotal?: Maybe<Scalars['Int']['output']>;
   createdAt: Scalars['BigInt']['output'];
@@ -2134,6 +2262,7 @@ export type Peer = {
   currentCCEndEpoch?: Maybe<Scalars['BigInt']['output']>;
   currentCCNextCCFailedEpoch?: Maybe<Scalars['BigInt']['output']>;
   currentCapacityCommitment?: Maybe<CapacityCommitment>;
+  deleted: Scalars['Boolean']['output'];
   /** ref to peerId in contract. */
   id: Scalars['ID']['output'];
   isAnyJoinedDeals: Scalars['Boolean']['output'];
@@ -2236,6 +2365,10 @@ export type Peer_Filter = {
   currentCapacityCommitment_not_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
   currentCapacityCommitment_starts_with?: InputMaybe<Scalars['String']['input']>;
   currentCapacityCommitment_starts_with_nocase?: InputMaybe<Scalars['String']['input']>;
+  deleted?: InputMaybe<Scalars['Boolean']['input']>;
+  deleted_in?: InputMaybe<Array<Scalars['Boolean']['input']>>;
+  deleted_not?: InputMaybe<Scalars['Boolean']['input']>;
+  deleted_not_in?: InputMaybe<Array<Scalars['Boolean']['input']>>;
   id?: InputMaybe<Scalars['ID']['input']>;
   id_gt?: InputMaybe<Scalars['ID']['input']>;
   id_gte?: InputMaybe<Scalars['ID']['input']>;
@@ -2324,6 +2457,7 @@ export type Peer_OrderBy =
   | 'currentCapacityCommitment__submittedProofsCount'
   | 'currentCapacityCommitment__totalCollateral'
   | 'currentCapacityCommitment__totalFailCount'
+  | 'deleted'
   | 'id'
   | 'isAnyJoinedDeals'
   | 'joinedDeals'
@@ -2347,6 +2481,7 @@ export type Peer_OrderBy =
 export type Provider = {
   __typename?: 'Provider';
   approved: Scalars['Boolean']['output'];
+  /** It depends on if CU in deal or not. */
   computeUnitsAvailable: Scalars['Int']['output'];
   computeUnitsTotal: Scalars['Int']['output'];
   createdAt: Scalars['BigInt']['output'];
@@ -2354,7 +2489,7 @@ export type Provider = {
   name: Scalars['String']['output'];
   offers?: Maybe<Array<Offer>>;
   peerCount: Scalars['Int']['output'];
-  /** Is provider registered in the network (if false - possible just mentioned). */
+  /** Is provider registered in the network (if false possibly it is only mentioned or global-whitelisted). */
   registered: Scalars['Boolean']['output'];
 };
 
@@ -2480,6 +2615,8 @@ export type Query = {
   deals: Array<Deal>;
   effector?: Maybe<Effector>;
   effectors: Array<Effector>;
+  epochStatistic?: Maybe<EpochStatistic>;
+  epochStatistics: Array<EpochStatistic>;
   graphNetwork?: Maybe<GraphNetwork>;
   graphNetworks: Array<GraphNetwork>;
   offer?: Maybe<Offer>;
@@ -2697,6 +2834,24 @@ export type QueryEffectorsArgs = {
   skip?: InputMaybe<Scalars['Int']['input']>;
   subgraphError?: _SubgraphErrorPolicy_;
   where?: InputMaybe<Effector_Filter>;
+};
+
+
+export type QueryEpochStatisticArgs = {
+  block?: InputMaybe<Block_Height>;
+  id: Scalars['ID']['input'];
+  subgraphError?: _SubgraphErrorPolicy_;
+};
+
+
+export type QueryEpochStatisticsArgs = {
+  block?: InputMaybe<Block_Height>;
+  first?: InputMaybe<Scalars['Int']['input']>;
+  orderBy?: InputMaybe<EpochStatistic_OrderBy>;
+  orderDirection?: InputMaybe<OrderDirection>;
+  skip?: InputMaybe<Scalars['Int']['input']>;
+  subgraphError?: _SubgraphErrorPolicy_;
+  where?: InputMaybe<EpochStatistic_Filter>;
 };
 
 
@@ -2989,8 +3144,6 @@ export type SubmittedProof_OrderBy =
   | 'capacityCommitment'
   | 'capacityCommitmentStatsPerEpoch'
   | 'capacityCommitmentStatsPerEpoch__activeUnitCount'
-  | 'capacityCommitmentStatsPerEpoch__blockNumberEnd'
-  | 'capacityCommitmentStatsPerEpoch__blockNumberStart'
   | 'capacityCommitmentStatsPerEpoch__computeUnitsWithMinRequiredProofsSubmittedCounter'
   | 'capacityCommitmentStatsPerEpoch__currentCCNextCCFailedEpoch'
   | 'capacityCommitmentStatsPerEpoch__epoch'
@@ -3022,6 +3175,7 @@ export type SubmittedProof_OrderBy =
   | 'capacityCommitment__totalFailCount'
   | 'computeUnit'
   | 'computeUnit__createdAt'
+  | 'computeUnit__deleted'
   | 'computeUnit__id'
   | 'computeUnit__submittedProofsCount'
   | 'computeUnit__workerId'
@@ -3035,6 +3189,7 @@ export type SubmittedProof_OrderBy =
   | 'peer__currentCCCollateralDepositedAt'
   | 'peer__currentCCEndEpoch'
   | 'peer__currentCCNextCCFailedEpoch'
+  | 'peer__deleted'
   | 'peer__id'
   | 'peer__isAnyJoinedDeals'
   | 'provider'
@@ -3073,6 +3228,8 @@ export type Subscription = {
   deals: Array<Deal>;
   effector?: Maybe<Effector>;
   effectors: Array<Effector>;
+  epochStatistic?: Maybe<EpochStatistic>;
+  epochStatistics: Array<EpochStatistic>;
   graphNetwork?: Maybe<GraphNetwork>;
   graphNetworks: Array<GraphNetwork>;
   offer?: Maybe<Offer>;
@@ -3290,6 +3447,24 @@ export type SubscriptionEffectorsArgs = {
   skip?: InputMaybe<Scalars['Int']['input']>;
   subgraphError?: _SubgraphErrorPolicy_;
   where?: InputMaybe<Effector_Filter>;
+};
+
+
+export type SubscriptionEpochStatisticArgs = {
+  block?: InputMaybe<Block_Height>;
+  id: Scalars['ID']['input'];
+  subgraphError?: _SubgraphErrorPolicy_;
+};
+
+
+export type SubscriptionEpochStatisticsArgs = {
+  block?: InputMaybe<Block_Height>;
+  first?: InputMaybe<Scalars['Int']['input']>;
+  orderBy?: InputMaybe<EpochStatistic_OrderBy>;
+  orderDirection?: InputMaybe<OrderDirection>;
+  skip?: InputMaybe<Scalars['Int']['input']>;
+  subgraphError?: _SubgraphErrorPolicy_;
+  where?: InputMaybe<EpochStatistic_Filter>;
 };